Fix WASM build
