+++ packages/dao-interface/Cargo.toml
@@ -15,6 +15,7 @@ cw20 = { workspace = true }
 cw721 = { workspace = true }
 cw-hooks = { workspace = true }
 cw-utils = { workspace = true }
+osmosis-std = { workspace = true }
 
 [dev-dependencies]
 cosmwasm-schema = { workspace = true }
