+++ packages/dao-interface/src/lib.rs
@@ -5,4 +5,5 @@ pub mod msg;
 pub mod proposal;
 pub mod query;
 pub mod state;
+pub mod token;
 pub mod voting;
