+++ test-contracts/dao-test-custom-factory/src/lib.rs
@@ -3,4 +3,3 @@ mod error;
 pub mod msg;
 
 pub use crate::error::ContractError;
-pub use dao_voting_token_staked::msg::{InitialBalance, NewTokenInfo};
